Moved phylorefs state code into PhylorefTable.
