+++ src/components/PhylorefTable.vue
@@ -74,6 +74,8 @@
  * and the ability to add new phyloreferences.
  */
 
+import { has, isEqual, uniq, chunk } from 'lodash';
+import Vue from 'vue';
 import { mapState } from 'vuex';
 
 export default {
@@ -81,9 +83,22 @@ export default {
   data: function () {
     return {
       flagDisplayExpression: false,
+      loadedPhylorefs: [],
+      openTreeTaxonomyInfoByName: {},
     };
   },
   computed: {
+    phylorefsWithMoreThanOneSpecifier() {
+      return this.loadedPhylorefs.filter(phyloref => (this.getSpecifiersForPhyloref(phyloref) || []).length > 1);
+    },
+    phylorefsWithEveryOTTIds() {
+      return this.loadedPhylorefs.filter(phyloref => {
+        const specifiers = this.getSpecifiersForPhyloref(phyloref);
+        if(specifiers.length > 0 &&
+          specifiers.length === specifiers.filter(specifier => this.getOpenTreeTaxonomyID(specifier) !== undefined).length
+        ) return true;
+      });
+    },
     exampleJSONLDURLs() { return [
       // Returns a list of example files to display in the "Examples" menu.
       {
@@ -111,15 +126,16 @@ export default {
       if(phyloref === undefined) return { internalTUs: [], externalTUs: [], allTUs: [] };
 
       // Returns a list of TUs for a particular phyloreference.
-      return this.$store.getters.getSpecifiersForPhyloref(phyloref);
+      return this.getSpecifiersForPhyloref(phyloref);
     },
+
     loadJSONLDFromURL(url) {
       // Change the current PHYX to that in the provided URL.
       // Will ask the user to confirm before replacing it.
 
       $.getJSON(url)
         .done((data) => {
-          this.$store.commit('extractPhyloreferencesFromJSONLD', data);
+          this.extractPhyloreferencesFromJSONLD(data);
         })
         .fail((error) => {
           if (error.status === 200) {
@@ -167,11 +183,209 @@ export default {
           const lines = e.target.result;
           const jsonld = JSON.parse(lines);
 
-          this.$store.commit('extractPhyloreferencesFromJSONLD', jsonld);
+          this.extractPhyloreferencesFromJSONLD(jsonld);
         });
         fr.readAsText(file);
       }
     },
+    addPhyloref(phyloref) {
+      // Check to make sure this phyloref hasn't already been added.
+      if(this.loadedPhylorefs.find(phy => isEqual(phy, phyloref)) !== undefined) return;
+
+      // No previous match? Then add it in!
+      this.loadedPhylorefs.push(phyloref);
+    },
+
+    extractPhyloreferencesFromJSONLD(jsonld) {
+      // Extract phyloreferences from the provided JSONLD file and add them to
+      // state.loaded. We use isEqual to prevent adding the same phyloreference
+      // more than once, but we will add different phyloreferences with the
+      // same '@id'.
+
+      // JSON-LD files sometimes contain an array of elements. In this case,
+      // we should try adding every one.
+      if(Array.isArray(jsonld)) {
+        jsonld.forEach(element => this.extractPhyloreferencesFromJSONLD(element));
+      }
+
+      // If this was generated by the Authoring Tool, then we can find phyloreferences
+      // just by looking for jsonld.phylorefs.
+      if(has(jsonld, 'phylorefs') && Array.isArray(jsonld.phylorefs)) {
+        jsonld.phylorefs.forEach(phy => this.addPhyloref(phy));
+      }
+
+      // If it was created by phyx2ontology, the phyloreferences can be recognized
+      // has having a subClassOf 'phyloref:Phyloreference'. Let's look for that.
+      if(has(jsonld, 'subClassOf')) {
+        if(Array.isArray(jsonld.subClassOf) && jsonld.subClassOf.includes('phyloref:Phyloreference'))
+          this.addPhyloref(jsonld);
+        if(jsonld.subClassOf === 'phyloref:Phyloreference')
+          this.addPhyloref(jsonld);
+      }
+    },
+
+    getSpecifiersFromClassExpression(classExpr) {
+      let results = [];
+
+      // If classExpr is an array, then process each entry separately.
+      if(Array.isArray(classExpr)) {
+        return classExpr
+          .map(expr => getSpecifiersFromClassExpression(expr))
+          .reduce((acc, val) => acc.concat(val), []);
+      }
+
+      // If classExpr itself has an equivalentClass, then we should process that instead.
+      if(has(classExpr, 'equivalentClass')) {
+        results = results.concat(getSpecifiersFromClassExpression(classExpr.equivalentClass));
+      }
+
+      // If there are additional classes, then process those too.
+      if(has(classExpr, 'hasAdditionalClass')) {
+        results = results.concat(getSpecifiersFromClassExpression(classExpr.hasAdditionalClass));
+      }
+
+      if(getLabelForSpecifier(classExpr) !== undefined) {
+        results.push(classExpr);
+      } else {
+        if(has(classExpr, 'someValuesFrom')) {
+          results = results.concat(getSpecifiersFromClassExpression(classExpr.someValuesFrom));
+        }
+
+        if(has(classExpr, 'intersectionOf')) {
+          results = results.concat(getSpecifiersFromClassExpression(classExpr.intersectionOf));
+        }
+      }
+
+      return results;
+    },
+
+    getLabelForSpecifier(expr) {
+      // Recognize the three standard expression forms.
+      // Form 1. obo:CDAO_0000149 some (excludes_TU some ...) and (includes_TU some ...)
+      if (
+        has(expr, 'onProperty') && expr.onProperty === 'obo:CDAO_0000149' &&
+        has(expr, 'someValuesFrom') && has(expr.someValuesFrom, 'intersectionOf')
+      ) {
+        // There are two possibilities here. We could be looking at the MRCA of two TUs,
+        // or we could be looking at the MRCA of one MRCA and one TU. So figure out which
+        // it is.
+        if(expr.someValuesFrom.intersectionOf.length === 2
+          && has(expr.someValuesFrom.intersectionOf[0], 'onProperty') && expr.someValuesFrom.intersectionOf[0].onProperty === 'phyloref:excludes_TU'
+          && has(expr.someValuesFrom.intersectionOf[1], 'onProperty') && expr.someValuesFrom.intersectionOf[1].onProperty === 'phyloref:includes_TU'
+        ) {
+          // Two-TU MRCA!
+          const comp1 = this.getLabelForSpecifierExpr(expr.someValuesFrom.intersectionOf[0]);
+          const comp2 = this.getLabelForSpecifierExpr(expr.someValuesFrom.intersectionOf[1]);
+
+          return `MRCA(${comp1.substr(25)}, ${comp2.substr(25)})`;
+        } else if(
+          expr.someValuesFrom.intersectionOf.length === 2
+          && has(expr.someValuesFrom.intersectionOf[0], 'onProperty') && expr.someValuesFrom.intersectionOf[0].onProperty === 'phyloref:excludes_lineage_to'
+        ) {
+          return undefined;
+        } else {
+          // No idea what this is!
+          return undefined;
+        }
+      }
+
+      // Form 2. includes_TU some (Name and scientificName some X)
+      else if (
+          (has(expr, 'onProperty') && expr.onProperty === 'phyloref:includes_TU') &&
+          (has(expr, 'someValuesFrom') && (
+            (has(expr.someValuesFrom, 'onProperty') && expr.someValuesFrom.onProperty === 'http://rs.tdwg.org/ontology/voc/TaxonConcept#hasName') &&
+            (has(expr.someValuesFrom, 'someValuesFrom') &&
+              (has(expr.someValuesFrom.someValuesFrom, 'intersectionOf')) &&
+              (has(expr.someValuesFrom.someValuesFrom.intersectionOf[1], 'onProperty') &&
+                expr.someValuesFrom.someValuesFrom.intersectionOf[1].onProperty === 'dwc:scientificName')
+            )
+          )
+        )
+      ) {
+          return `includes scientific name ${expr.someValuesFrom.someValuesFrom.intersectionOf[1].hasValue}`;
+          // return JSON.stringify(expr);
+      }
+      // Form 3. excludes_TU some (Name and scientificName some X)
+      else if (
+          (has(expr, 'onProperty') && expr.onProperty === 'phyloref:excludes_TU') &&
+          (has(expr, 'someValuesFrom') && (
+              (has(expr.someValuesFrom, 'onProperty') && expr.someValuesFrom.onProperty === 'http://rs.tdwg.org/ontology/voc/TaxonConcept#hasName') &&
+              (has(expr.someValuesFrom, 'someValuesFrom') &&
+                (has(expr.someValuesFrom.someValuesFrom, 'intersectionOf')) &&
+                (has(expr.someValuesFrom.someValuesFrom.intersectionOf[1], 'onProperty') &&
+                  expr.someValuesFrom.someValuesFrom.intersectionOf[1].onProperty === 'dwc:scientificName')
+              )
+          ))
+      ) {
+          return `excludes scientific name ${expr.someValuesFrom.someValuesFrom.intersectionOf[1].hasValue}`;
+          // return JSON.stringify(expr);
+      } else {
+          // Could not match!
+          return undefined;
+      }
+    },
+
+    convertTaxonomicUnitToExpr(type) {
+      return (tu) => {
+        const property = `phyloref:${type}_TU`;
+
+        if(!has(tu, 'referencesTaxonomicUnits')) return undefined;
+        // Only use the first taxonomic unit!
+        const tunit = tu.referencesTaxonomicUnits[0];
+        // Scientific name?
+        if (has(tunit, 'scientificNames')) {
+          return {
+            onProperty: property,
+            someValuesFrom: {
+              onProperty: 'http://rs.tdwg.org/ontology/voc/TaxonConcept#hasName',
+              someValuesFrom: {
+                intersectionOf: [
+                  {
+                    "@id": "obo:NOMEN_0000107"
+                  },
+                  {
+                    onProperty: 'dwc:scientificName',
+                    hasValue: tunit.scientificNames[0].scientificName
+                  }
+                ]
+              }
+            }
+          };
+        } else {
+          // We don't support anything else!
+          return undefined;
+        }
+      }
+    },
+
+    getSpecifiersForPhyloref(phyloref) {
+      // phyloref: Phyloreference to retrieve specifiers from.
+      //
+      // All phyloref objects should have internal and external specifier
+      // information. But first, let's see if we can extract it directly from
+      // the equivalentClass statement.
+
+      if(has(phyloref, 'internalSpecifiers')) {
+        // Old form! Let's just extract the taxonomic units and go with that.
+        const internals = (phyloref['internalSpecifiers'] || [])
+          .map(this.convertTaxonomicUnitToExpr('includes'))
+          .filter(tu => tu !== undefined);
+        const externals = (phyloref['externalSpecifiers'] || [])
+          .map(this.convertTaxonomicUnitToExpr('excludes'))
+          .filter(tu => tu !== undefined);
+
+        if(internals.length === 0) return [];
+        if(externals.length === 0) {
+          return internals;
+        }
+        return internals.concat(externals);
+      }
+
+      return uniqWith(
+        this.getSpecifiersFromClassExpression(phyloref || []),
+        isEqual
+      );
+    }
   }
 };
 </script>
