+++ src/store/index.js
@@ -2,9 +2,6 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 
-// Load modules
-import phylorefs from './modules/phylorefs';
-
 Vue.use(Vuex);
 
 const debug = process.env.NODE_ENV !== 'production';
@@ -12,7 +9,6 @@ const debug = process.env.NODE_ENV !== 'production';
 export default new Vuex.Store({
   state: {},
   modules: {
-    phylorefs,
   },
   strict: debug,
 });
